remove duplicate code
