+++ test/openshift/cluster.test.ts
@@ -17,6 +17,7 @@ import { TestItem } from './testOSItem';
 import { OpenShiftItem } from '../../src/openshift/openshiftItem';
 import pq = require('proxyquire');
 import { getVscodeModule } from '../../src/util/credentialManager';
+import { HidePassword } from '../../src/util/hidepassword';
 
 const expect = chai.expect;
 chai.use(sinonChai);
@@ -88,7 +89,7 @@ suite('Openshift/Cluster', () => {
             quickPickStub.onThirdCall().resolves({description: "Current Context", label: testUser});
             inputStub.resolves(password);
             commandStub.rejects(error);
-            sandbox.stub(Cluster, 'hideTokenpassword').returns('FATAL ERROR');
+            sandbox.stub(HidePassword, 'hideTokenpassword').returns('FATAL ERROR');
 
             try {
                 await Cluster.login();
@@ -163,7 +164,7 @@ suite('Openshift/Cluster', () => {
 
             test('errors if there is output on odo stderr', async () => {
                 execStub.resolves(errorData);
-                sandbox.stub(Cluster, 'hideCredentialPassword').returns('FATAL ERROR');
+                sandbox.stub(HidePassword, 'hideCredentialPassword').returns('FATAL ERROR');
 
                 try {
                     await Cluster.credentialsLogin();
@@ -249,7 +250,7 @@ suite('Openshift/Cluster', () => {
 
             test('handles incoming errors the same way as credentials login', async () => {
                 execStub.rejects(error);
-                sandbox.stub(Cluster, 'hideTokenpassword').returns('FATAL ERROR');
+                sandbox.stub(HidePassword, 'hideTokenpassword').returns('FATAL ERROR');
                 try {
                     await Cluster.tokenLogin();
                     expect.fail();
@@ -315,7 +316,7 @@ suite('Openshift/Cluster', () => {
         });
 
         test('throws errors from subsequent login', async () => {
-            sandbox.stub(Cluster, 'hideTokenpassword').returns('FATAL ERROR');
+            sandbox.stub(HidePassword, 'hideTokenpassword').returns('FATAL ERROR');
             execStub.onSecondCall().resolves(errorData);
             infoStub.resolves('Yes');
             quickPickStub.onFirstCall().resolves('Credentials');
